refactor: 배달 상태 검증 코드 추가 및 테스트코드 수정

+++ src/main/java/com/project/food_ordering_service/domain/delivery/service/DeliveryService.java
@@ -92,6 +92,11 @@ public class DeliveryService {
                 .orElseThrow(DeliveryNotFoundException::new);
 
         Order order = delivery.getOrder();
+
+        if (order.getStatus() != OrderStatus.RECEIVED) {
+            throw new IllegalStateException("배달 상태가 RECEIVED일 때만 배달을 취소할 수 있습니다.");
+        }
+
         order.updateOrderStatus(OrderStatus.DELIVERY_REQUESTED);
 
         deliveryRepository.delete(delivery);
